# Backend/movie-app-go/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Genre:
    properties:
      genre_id:
        type: string
      genre_name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - genre_id
    - genre_name
    type: object
  models.LoginResponse:
    properties:
      access_token:
        type: string
      email:
        type: string
      favourite_movies_genres:
        items:
          $ref: '#/definitions/models.Genre'
        type: array
      first_name:
        type: string
      last_name:
        type: string
      refresh_token:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
...
  models.UpdateUser:
    properties:
      email:
        type: string
      favourite_movies_genres:
        items:
          type: string
        type: array
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
  models.User:
    properties:
      access_token:
        type: string
      created_at:
        type: string
      email:
        type: string
      favourite_movies_genres:
        items:
          $ref: '#/definitions/models.Genre'
        type: array
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      refresh_token:
        type: string
      role:
        enum:
        - ADMIN
        - USER
        - GUEST
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    required:
    - email
    - favourite_movies_genres
    - first_name
    - last_name
    - password
    type: object
  models.UserLogin:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  models.UserLogout:
    properties:
      user_id:
        type: string
    type: object
host: localhost:5000
info:
  contact: {}
  description: API for the movie app.
  title: Movie API
  version: "1.0"
paths:
...
  /api/v1/deleteuser/{userId}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete user
      tags:
      - users
  /api/v1/getuserbyID/{userId}:
    get:
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get user by id
      tags:
      - users
  /api/v1/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Login
      tags:
      - auth
  /api/v1/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Logout
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UserLogout'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Logout
      tags:
      - auth
  /api/v1/refresh-token:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Refresh tokens
      tags:
      - auth
  /api/v1/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: User
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - auth
  /api/v1/updateuser/{userId}:
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Updates
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - users
  /api/v1/users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: List users
      tags:
      - users
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"